Add student assets

+++ src/main/java/es/upm/tfgback/controller/TFGController.java
@@ -1,6 +1,5 @@
 package es.upm.tfgback.controller;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -9,15 +8,7 @@ import es.upm.tfgback.repository.TFGRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api")
@@ -26,15 +17,24 @@ public class TFGController {
     @Autowired
     TFGRepository tfgRepository;
 
+    @CrossOrigin
     @GetMapping("/tfgs")
-    public ResponseEntity<List<TFG>> getAllTFGs(@RequestParam(required = false) String title) {
+    public ResponseEntity<List<TFG>> getAllTFGs() {
         try {
-            List<TFG> tfgs = new ArrayList<TFG>();
-            if (title == null)
-                tfgs.addAll(tfgRepository.findAll());
-            else {
-                tfgs.addAll(tfgRepository.findByTitleContaining(title));
+            List<TFG> tfgs = tfgRepository.findAll();
+            if (tfgs.isEmpty()) {
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
             }
+            return new ResponseEntity<>(tfgs, HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @GetMapping("/tfgs/students/{studentId}")
+    public ResponseEntity<List<TFG>> getTFGByAuthor(@PathVariable("studentId") long studentId) {
+        try {
+            List<TFG> tfgs = tfgRepository.findByStudentId(studentId);
             if (tfgs.isEmpty()) {
                 return new ResponseEntity<>(HttpStatus.NO_CONTENT);
             }
@@ -51,10 +51,10 @@ public class TFGController {
                 .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
-    @PostMapping("/tfgs") // TODO no autoincrementa la secuencia
+    @PostMapping("/tfgs")
     public ResponseEntity<TFG> createTFG(@RequestBody TFG tfg) {
         try {
-            TFG _tfg = tfgRepository.save(new TFG(tfg.getTitle()));
+            TFG _tfg = tfgRepository.save(new TFG(tfg.getStudentId(), tfg.getTitle()));
             return new ResponseEntity<>(_tfg, HttpStatus.CREATED);
         } catch (Exception e) {
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
@@ -76,8 +76,13 @@ public class TFGController {
     @DeleteMapping("/tfgs/{id}")
     public ResponseEntity<HttpStatus> deleteTFG(@PathVariable("id") long id) {
         try {
-            tfgRepository.deleteById(id);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            Optional<TFG> tfgData = tfgRepository.findById(id);
+            if (tfgData.isPresent()) {
+                tfgRepository.deleteById(id);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
